Removed extra demographic model object #2

+++ src/main/java/oscar/billing/controler/ConsultaFaturamentoPacienteAction.java
@@ -1,25 +1,25 @@
 /*
- * 
+ *
  * Copyright (c) 2001-2002. Department of Family Medicine, McMaster University. All Rights Reserved. *
- * This software is published under the GPL GNU General Public License. 
- * This program is free software; you can redistribute it and/or 
- * modify it under the terms of the GNU General Public License 
- * as published by the Free Software Foundation; either version 2 
- * of the License, or (at your option) any later version. * 
- * This program is distributed in the hope that it will be useful, 
- * but WITHOUT ANY WARRANTY; without even the implied warranty of 
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
- * GNU General Public License for more details. * * You should have received a copy of the GNU General Public License 
- * along with this program; if not, write to the Free Software 
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA. * 
- * 
+ * This software is published under the GPL GNU General Public License.
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version. *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details. * * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA. *
+ *
  * <OSCAR TEAM>
- * 
- * This software was written for the 
- * Department of Family Medicine 
- * McMaster University 
- * Hamilton 
- * Ontario, Canada 
+ *
+ * This software was written for the
+ * Department of Family Medicine
+ * McMaster University
+ * Hamilton
+ * Ontario, Canada
  */
 package oscar.billing.controler;
 
@@ -30,10 +30,10 @@ import org.apache.log4j.Logger;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
+import org.oscarehr.common.model.Demographic;
 import org.oscarehr.util.MiscUtils;
 
 import oscar.billing.dao.AppointmentDAO;
-import oscar.billing.model.Demographic;
 import oscar.util.OscarAction;
 
 
@@ -74,11 +74,11 @@ public class ConsultaFaturamentoPacienteAction extends OscarAction {
         try {
             form.clear();
             logger.info(" [ConsultaFaturamentoPacienteAction] Limpou form");
-            
+
             String id = request.getParameter("demographic_no");
 			logger.info(" [ConsultaFaturamentoPacienteAction] id " + id);
             Demographic demographic = new Demographic();
-            demographic.setDemographicNo(Long.parseLong(id));
+            demographic.setDemographicNo(Integer.parseInt(id));
 
             AppointmentDAO appDAO = new AppointmentDAO();
